# Finalize some classes