--- a/src/passes/DeNaN.cpp
+++ b/src/passes/DeNaN.cpp
@@ -30,8 +30,9 @@
 
 namespace wasm {
 
-struct DeNaN : public WalkerPass<
-                 ControlFlowWalker<DeNaN, UnifiedExpressionVisitor<DeNaN>>> {
+struct DeNaN final
+  : public WalkerPass<
+      ControlFlowWalker<DeNaN, UnifiedExpressionVisitor<DeNaN>>> {
 
   Name deNan32, deNan64;