… src/passes/GlobalRefining.cpp
… src/passes/GlobalStructInference.cpp
… src/passes/Heap2Local.cpp
… src/passes/I64ToI32Lowering.cpp
… src/passes/Inlining.cpp
… src/passes/InstrumentLocals.cpp
… src/passes/InstrumentMemory.cpp
… src/passes/Intrinsics.cpp
… src/passes/LegalizeJSInterface.cpp
… src/passes/LimitSegments.cpp
… src/passes/LocalCSE.cpp
… src/passes/LocalSubtyping.cpp
… src/passes/LogExecution.cpp
… src/passes/LoopInvariantCodeMotion.cpp
… src/passes/Memory64Lowering.cpp
… src/passes/MemoryPacking.cpp
… src/passes/MergeBlocks.cpp
… src/passes/MergeLocals.cpp
… src/passes/MergeSimilarFunctions.cpp
… src/passes/Metrics.cpp
… src/passes/NameList.cpp
… src/passes/NameTypes.cpp
… src/passes/OnceReduction.cpp
… src/passes/OptimizeAddedConstants.cpp
… src/passes/OptimizeForJS.cpp
… src/passes/OptimizeInstructions.cpp
… src/passes/PickLoadSigns.cpp
… src/passes/Poppify.cpp
… src/passes/PostEmscripten.cpp
… src/passes/Precompute.cpp