This PR introduces minor optimization for classes. Adding "final" for derived classes with virtual members may reduce dynamic dispatching