@tlively 2022-06-21
I don't see any reason _not_ to merge this, but I would be surprised if it is useful. Nearly all these classes use the curiously recursive template pattern to avoid dynamic dispatch and the classes don't even contain virtual methods, so I don't think there are any additional optimizations that `final` would allow. Out of curiosity, @MaxGraey, have you been able to measure a performance difference due to this PR?