@MaxGraey 2022-06-22
Sure! I checked perf. This results:

Bench results for 
```bash
BINARYEN_PASS_DEBUG=1 ./wasm-opt in.wasm -o out.wasm -O4
```

[in.wasm.zip](https://github.com/WebAssembly/binaryen/files/8956538/in.wasm.zip) (1.3 MB)

master (total: 8.26331 seconds)
this PR (total: 7.40632 seconds)

master (detailed):
```
[PassRunner]   running pass: duplicate-function-elimination...    0.00447184 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: memory-packing...                    9.1436e-05 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: once-reduction...                    0.00304166 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: ssa-nomerge...                       0.219998 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: flatten...                           0.459585 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-locals-notee-nostructure... 0.536922 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: local-cse...                         0.183286 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: dce...                               0.17535 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-names...               0.0418942 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-brs...                 0.164733 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-names...               0.0419508 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: optimize-instructions...             0.150581 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: pick-load-signs...                   0.0666631 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: precompute-propagate...              0.458745 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: code-pushing...                      0.0781121 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-locals-nostructure...       0.475734 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.298806 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: reorder-locals...                    0.0855048 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-brs...                 0.103497 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: merge-locals...                      0.290443 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: coalesce-locals...                   0.252365 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: local-cse...                         0.101981 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-locals...                   0.393886 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.145751 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: reorder-locals...                    0.0364165 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: coalesce-locals...                   0.188113 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: reorder-locals...                    0.0374981 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.146023 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: code-folding...                      0.116423 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: merge-blocks...                      0.0833095 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-brs...                 0.0960227 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-names...               0.0292915 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: merge-blocks...                      0.0806398 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: precompute-propagate...              0.365993 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: optimize-instructions...             0.117615 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: rse...                               0.154563 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.143864 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: dae-optimizing...                    0.0222761 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: inlining-optimizing...               1.33781 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: duplicate-function-elimination...    0.00391867 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: duplicate-import-elimination...      9.4478e-05 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-globals-optimizing...       0.0155929 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-module-elements...     0.02909 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: directize...                         0.00328307 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: generate-stack-ir...                 0.0319071 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: optimize-stack-ir...                 0.490175 seconds.
[PassRunner]   (validating)
[PassRunner] passes took 8.26331 seconds.
```

this PR (detailed)
```
[PassRunner] running passes
[PassRunner]   running pass: duplicate-function-elimination...    0.00422479 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: memory-packing...                    9.0594e-05 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: once-reduction...                    0.00331366 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: ssa-nomerge...                       0.224965 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: flatten...                           0.374685 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-locals-notee-nostructure... 0.473384 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: local-cse...                         0.163811 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: dce...                               0.166299 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-names...               0.0347605 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-brs...                 0.136641 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-names...               0.0399032 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: optimize-instructions...             0.132677 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: pick-load-signs...                   0.0471047 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: precompute-propagate...              0.446167 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: code-pushing...                      0.064864 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-locals-nostructure...       0.413889 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.279204 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: reorder-locals...                    0.0778189 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-brs...                 0.0874164 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: merge-locals...                      0.268711 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: coalesce-locals...                   0.243312 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: local-cse...                         0.103932 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-locals...                   0.332937 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.134011 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: reorder-locals...                    0.0297158 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: coalesce-locals...                   0.191064 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: reorder-locals...                    0.0287478 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.126502 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: code-folding...                      0.109245 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: merge-blocks...                      0.0674473 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-brs...                 0.0824514 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-names...               0.0246729 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: merge-blocks...                      0.0638275 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: precompute-propagate...              0.331777 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: optimize-instructions...             0.108771 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: rse...                               0.152219 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: vacuum...                            0.137648 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: dae-optimizing...                    0.0212426 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: inlining-optimizing...               1.20332 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: duplicate-function-elimination...    0.00413571 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: duplicate-import-elimination...      9.7026e-05 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: simplify-globals-optimizing...       0.0149864 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: remove-unused-module-elements...     0.0201101 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: directize...                         0.00299551 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: generate-stack-ir...                 0.0310583 seconds.
[PassRunner]   (validating)
[PassRunner]   running pass: optimize-stack-ir...                 0.400173 seconds.
[PassRunner]   (validating)
[PassRunner] passes took 7.40632 seconds.
```